refactor(extractors): rename scripts/walkers/ → scripts/extractors/
"Walker" applies precisely to AST traversal; the per-engine modules
combine introspection + AST walking. Umbrella term should be
"extractors". This rename uses git mv to preserve history.

Renames:
- scripts/walkers/ → scripts/extractors/
- per-engine modules get _extractor suffix:
  - tensorrt_ast.py → tensorrt_ast_extractor.py
  - tensorrt_introspection.py → tensorrt_introspection_extractor.py
  - vllm_ast.py → vllm_ast_extractor.py
  - vllm_introspection.py → vllm_introspection_extractor.py
  - transformers_ast.py → transformers_ast_extractor.py
  - transformers_introspection.py → transformers_introspection_extractor.py
  - transformers.py → transformers_extractor.py
- tests/unit/scripts/walkers/ → tests/unit/scripts/extractors/
- tests/unit/walkers/ → tests/unit/extractors/
- corresponding test files renamed with _extractor suffix

Updates all imports across:
- src/ (loader.py, library_resolution.py)
- tests/ (integration + unit)
- .github/workflows/config-rules-refresh.yml
- scripts/update_engine_rules.sh
- configs/validation_rules/README.md
- scripts/extractors/build_corpus.py registry

Schema field names preserved (WalkerSource, walker_source, walker_confidence)
since they're part of the YAML corpus schema, not internal naming.

Also fixes pre-existing tensorrt_ast lint debt (parallel to #417 vllm fix):
- Add noqa: E402 on yaml import
- Remove unused 'msg' variable in visit_Assert
- Drop unused tensorrt_llm config imports (KvCacheConfig, SchedulerConfig,
  QuantConfig) — only LlmConfig is actually used; keep tensorrt_llm import
  with noqa for availability check.
