+++ .github/workflows/config-rules-refresh.yml
@@ -12,7 +12,7 @@ name: Config Rules Refresh
 #   - docker/Dockerfile.{engine}        (Renovate-driven image bumps)
 #   - configs/validation_rules/{engine}.yaml
 #   - scripts/vendor_rules.py
-#   - scripts/walkers/{engine}.py
+#   - scripts/extractors/{engine}.py
 #   - scripts/_vendor_common.py
 #
 # Trigger (manual): workflow_dispatch with explicit engine choice.
@@ -25,7 +25,7 @@ on:
       - "configs/validation_rules/*.yaml"
       - "scripts/vendor_rules.py"
       - "scripts/_vendor_common.py"
-      - "scripts/walkers/*.py"
+      - "scripts/extractors/*.py"
       - "scripts/update_engine_rules.sh"
   workflow_dispatch:
     inputs:
@@ -110,7 +110,7 @@ jobs:
         run: |
           uv run python -c "
           import yaml
-          from scripts.walkers._fixpoint_test import fixpoint_test_corpus
+          from scripts.extractors._fixpoint_test import fixpoint_test_corpus
           corpus = yaml.safe_load(open('${CORPUS_DIR}/transformers.yaml'))
           fixpoint_test_corpus(corpus)
           print('fixpoint test passed')
@@ -131,10 +131,10 @@ jobs:
             "configs/validation_rules/transformers.yaml"
             "scripts/vendor_rules.py"
             "scripts/_vendor_common.py"
-            "scripts/walkers/transformers.py"
-            "scripts/walkers/transformers_introspection.py"
-            "scripts/walkers/_fixpoint_test.py"
-            "scripts/walkers/_base.py"
+            "scripts/extractors/transformers_extractor.py"
+            "scripts/extractors/transformers_introspection_extractor.py"
+            "scripts/extractors/_fixpoint_test.py"
+            "scripts/extractors/_base.py"
             "scripts/update_engine_rules.sh"
           )
           SHA=$(git log -1 --format=%H -- "${PATHS[@]}")
@@ -275,7 +275,7 @@ jobs:
         run: |
           uv run python -c "
           import yaml
-          from scripts.walkers._fixpoint_test import fixpoint_test_corpus
+          from scripts.extractors._fixpoint_test import fixpoint_test_corpus
           corpus = yaml.safe_load(open('${CORPUS_DIR}/vllm.yaml'))
           fixpoint_test_corpus(corpus)
           print('fixpoint test passed')
@@ -289,10 +289,10 @@ jobs:
             "configs/validation_rules/vllm.yaml"
             "scripts/vendor_rules.py"
             "scripts/_vendor_common.py"
-            "scripts/walkers/vllm_ast.py"
-            "scripts/walkers/vllm_introspection.py"
-            "scripts/walkers/_fixpoint_test.py"
-            "scripts/walkers/_base.py"
+            "scripts/extractors/vllm_ast_extractor.py"
+            "scripts/extractors/vllm_introspection_extractor.py"
+            "scripts/extractors/_fixpoint_test.py"
+            "scripts/extractors/_base.py"
           )
           SHA=$(git log -1 --format=%H -- "${PATHS[@]}")
           DATE=$(git log -1 --format=%aI -- "${PATHS[@]}")
