+++ scripts/extractors/__init__.py
@@ -1,6 +1,6 @@
 """Walkers that extract validation rules from engine library source.
 
-Each ``scripts/walkers/{engine}.py`` walker (added per-engine in follow-up
+Each ``scripts/extractors/{engine}.py`` walker (added per-engine in follow-up
 PRs) is version-pinned to a specific library release via the walker module's
 ``TESTED_AGAINST_VERSIONS`` and emits a corpus-compatible YAML document of
 rule candidates.
@@ -13,7 +13,7 @@ Two extraction mechanisms are in scope:
   call and formats the results.
 - **AST source parsing** — when no such API exists (vLLM, TRT-LLM), the
   walker parses the library source AST using the primitives in
-  :mod:`scripts.walkers._base`.
+  :mod:`scripts.extractors._base`.
 
 No concrete walker ships in this module today; they land as independent PRs
 per engine.
