+++ scripts/extractors/vllm_ast_extractor.py
@@ -4,7 +4,7 @@ Walks ``SamplingParams.__post_init__`` source to detect silent assignments
 (dormancy) and warned assignments (announced dormancy) patterns.
 
 Output schema: ``configs/validation_rules/_staging/vllm_ast.yaml``
-consumed downstream by ``scripts/walkers/build_corpus.py``.
+consumed downstream by ``scripts/extractors/build_corpus.py``.
 """
 
 from __future__ import annotations
@@ -28,7 +28,7 @@ _SCRIPT_DIR = str(Path(__file__).resolve().parent)
 sys.path[:] = [p for p in sys.path if Path(p).resolve() != Path(_SCRIPT_DIR).resolve()]
 sys.path[:] = [p for p in sys.path if p != ""]
 
-from scripts.walkers._base import (  # noqa: E402
+from scripts.extractors._base import (  # noqa: E402
     RuleCandidate,
     WalkerSource,
     call_func_path,
