+++ scripts/update_engine_rules.sh
@@ -71,7 +71,7 @@ OUTPUT_REL="src/llenergymeasure/config/vendored_rules/${ENGINE}.json"
 
 if [[ ! -f "$REPO_ROOT/$CORPUS_REL" ]]; then
     echo "[$ENGINE] Corpus $CORPUS_REL not found. Run the walker first:" >&2
-    echo "    python -m scripts.walkers.${ENGINE} --out $CORPUS_REL" >&2
+    echo "    python -m scripts.extractors.${ENGINE} --out $CORPUS_REL" >&2
     exit 1
 fi
