+++ tests/unit/config/vendored_rules/test_corpus_invariants.py
@@ -26,7 +26,7 @@ from llenergymeasure.config.vendored_rules import (
 )
 
 _VALID_CONFIDENCE = frozenset({"high", "medium", "low"})
-# SSOT: scripts/walkers/_base.py::Confidence / VALID_CONFIDENCE. Duplicated
+# SSOT: scripts/extractors/_base.py::Confidence / VALID_CONFIDENCE. Duplicated
 # here to avoid a cross-layer (engines → scripts) test import.
 
 
@@ -52,7 +52,7 @@ def test_corpus_covers_required_invariants(transformers_corpus) -> None:
 
     # Single-field invariants — at least one rule must touch each path.
     # Scope reflects the regenerated canonical corpus produced by
-    # ``scripts/walkers/build_corpus.py`` with the vendor-validation gate.
+    # ``scripts/extractors/build_corpus.py`` with the vendor-validation gate.
     # Adding a path here means: a rule for that field must survive vendor
     # validation against the pinned engine library (see ``engine_version``
     # in the corpus envelope). If a path drops out, investigate WHY (real
@@ -133,7 +133,7 @@ def test_corpus_covers_required_invariants(transformers_corpus) -> None:
     manual = [rule.id for rule in rules if rule.added_by == "manual_seed"]
     assert not manual, (
         f"corpus contains {len(manual)} hand-seeded rules; corpus must be "
-        f"machine-extracted (run scripts/walkers/build_corpus.py): {manual}"
+        f"machine-extracted (run scripts/extractors/build_corpus.py): {manual}"
     )
